+++ packages/bundler/src/polyfills/crypto/context/crypto.context.js
@@ -1,0 +1,68 @@
+/* eslint-disable */
+import * as crypto from 'node:crypto';
+
+export var { Cipher } = crypto;
+export var { Decipher } = crypto;
+export var { DiffieHellman } = crypto;
+export var { DiffieHellmanGroup } = crypto;
+export var { Hash } = crypto;
+export var { Hmac } = crypto;
+export var { Sign } = crypto;
+export var { Verify } = crypto;
+export var { constants } = crypto;
+export var { createCipheriv } = crypto;
+export var { createDecipheriv } = crypto;
+export var { createDiffieHellman } = crypto;
+export var { createDiffieHellmanGroup } = crypto;
+export var { createECDH } = crypto;
+export var { createHash } = crypto;
+export var { createHmac } = crypto;
+export var { createSign } = crypto;
+export var { createVerify } = crypto;
+export var { getCiphers } = crypto;
+export var { getDiffieHellman } = crypto;
+export var { getHashes } = crypto;
+export var { pbkdf2 } = crypto;
+export var { pbkdf2Sync } = crypto;
+export var { privateDecrypt } = crypto;
+export var { privateEncrypt } = crypto;
+export var { pseudoRandomBytes } = crypto;
+export var { publicDecrypt } = crypto;
+export var { publicEncrypt } = crypto;
+export var { randomBytes } = crypto;
+export var { randomFill } = crypto;
+export var { randomFillSync } = crypto;
+
+export default {
+  Cipher,
+  Decipher,
+  DiffieHellman,
+  DiffieHellmanGroup,
+  Hash,
+  Hmac,
+  Sign,
+  Verify,
+  constants,
+  createCipheriv,
+  createDecipheriv,
+  createDiffieHellman,
+  createDiffieHellmanGroup,
+  createECDH,
+  createHash,
+  createHmac,
+  createSign,
+  createVerify,
+  getCiphers,
+  getDiffieHellman,
+  getHashes,
+  pbkdf2,
+  pbkdf2Sync,
+  privateDecrypt,
+  privateEncrypt,
+  pseudoRandomBytes,
+  publicDecrypt,
+  publicEncrypt,
+  randomBytes,
+  randomFill,
+  randomFillSync,
+};