+++ packages/bundler/src/polyfills/crypto/context/index.js
@@ -1,0 +1,3 @@
+import cryptoContext from './crypto.context.js';
+
+export default cryptoContext;