+++ packages/bundler/src/polyfills/promises/context/index.js
@@ -1,0 +1,3 @@
+import promisesContext from './promises.context.js';
+
+export default promisesContext;