+++ packages/unenv-preset/src/polyfills/node/empty.js
@@ -1,0 +1,3 @@
+export function loadEnvConfig() {}
+
+export default {};