+++ packages/unenv-preset/src/polyfills/node/globals/process.cjs
@@ -154,8 +154,25 @@ Item.prototype.run = function () {
 };
 processShim.title = 'browser';
 processShim.browser = true;
-processShim.env = processShim.env =
-  typeof globalThis.process !== 'undefined' && globalThis.process.env ? globalThis.process.env : {};
+const initialValues = globalThis.process !== 'undefined' && globalThis.process?.env ? globalThis.process.env : {};
+Object.defineProperty(processShim, 'env', {
+  value: new Proxy(
+    { ...initialValues },
+    {
+      get(target, prop) {
+        return target[prop] || undefined;
+      },
+      set(target, prop, value) {
+        target[prop] = value;
+        return true;
+      },
+    },
+  ),
+  writable: true,
+  enumerable: true,
+  configurable: true,
+});
+
 processShim.argv = [];
 processShim.version = ''; // empty string to avoid regexp issues
 processShim.versions = { node: '18.3.1' };
@@ -185,11 +202,12 @@ processShim.cwd = function () {
 };
 
 processShim.chdir = function (dir) {
-  throw new Error('process.chdir is not supported');
+  return dir ?? '/';
 };
 processShim.umask = function () {
   return 0;
 };
 
 globalThis.process = processShim;
-globalThis.process.env = processShim.env;
+
+module.exports = processShim;